backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
python-dotenv
pydantic-settings
huggingface-hub
websockets
twilio